perf(train): optimize rtx3070 run + fix compile/binary-factorized crashes
Tune the canonical config for RTX 3070 (8GB, mem_reserved <=7.4GB) and fix
two crashes that blocked training under manual_bf16 + torch.compile.

Config (rtx3070_canonical.yaml):
- msa_slot_count 2048->4096: doubles generation memory context 8192->16384
  tokens. Training-neutral: hagi.py:576 clears the model registry every
  forward, so slot_count is only the eviction cap (never hit in training);
  the larger cap only helps generation (chat.py:59 persistent registry).
- msa_chunk_size 4 (kept): 4->8 profiled at reserved 9.54GB (>7.4 limit) and
  tokens/sec 7253 via quadratic MSA attention growth — reverted.
- msa_top_k 5->8, hdim_heads 4->6, label_smoothing 0.1->0.05, rope_theta
  10000->500000, transformer.max_seq_len 2048->4096, min_seq_len 16->256,
  ce_chunk_size/ce_fused 2048->4096, log_interval 25->50, loss_warmup
  linear->cosine, muon_ns_steps 5->6, aux_warmup 2000->1500, iso_warmup
  3000->2000, lsh disabled (Python-loop fwd killer), gradient_checkpointing on.

Code:
- binary_factorized.py: torch.where(1.0,-1.0) promoted binarized matrix to
  fp32 -> F.linear(x_bf16, w_fp32) RuntimeError under manual_bf16 (no
  autocast). Cast result to x.dtype; STE semantics preserved.
- msa.py: @torch.compiler.disable on SlotRegistry methods. Dynamo was
  specializing a guard on len(self._slot_ids) (changes every forward:
  clear->0->fill->N) -> recompile_limit(8) -> eager fallback (step-0 fwd
  319s wasted compile). Registry is stateful Python bookkeeping, not a
  differentiable module; tensor ops (matmul/topk) stay compiled.
- loop.py: drop w_aux/w_iso/h=/l= from the step log line (weights/cycles
  are config, not per-step signal). Keep effective_weights/hrm computation.
- train.py: set PYTORCH_CUDA_ALLOC_CONF=expandable_segments before torch
  import (Windows build ignores it but reserves stay bounded via GC).

Profiled 30 steps each: reserved 5.25-6.46GB (<7.4), tokens/sec 17291
(no compile) / ~22000 (compile), fwd 645ms->598ms, loss 10.97->8.09@75.

Constraint: mem_reserved <= 7.4GB; only one training/profile process at a time
Rejected: msa_chunk_size 8 | reserved 9.54GB > limit, tokens/sec -58%
Rejected: moe_top_k 2 | tokens/sec 17291->11989 (-31%), slow-path atomics
Rejected: ce_chunk_size 8192 | negligible speed, +1.5-2.3GB VRAM-peak risk
Rejected: hdim_delay_steps 2 | inert at hrm_l_cycles=1 (no delay budget)
Directive: do NOT raise msa_chunk_size past 4 without re-profiling reserved;
  the LSH path (msa_lsh_threshold>0) stays disabled — its per-query Python
  loop dominates fwd at Q=B*T=8192.
Confidence: high
Scope-risk: moderate
Not-tested: full multi-epoch convergence (profiles were 30-75 steps); binary
  factorization only smoke-tested for forward correctness, not quality

+++ src/hagi/model/binary_factorized.py
@@ -38,7 +38,13 @@ class BinaryFactorizedLinear(nn.Module):
 
         Forward uses sign(x). Backward flows as if x itself.
         """
-        b = torch.where(x >= 0, 1.0, -1.0)
+        # torch.where with Python float scalars promotes to fp32 (torch default
+        # dtype), which breaks under manual_bf16 precision (no autocast): the
+        # model is cast to bf16 (loop.py:431) but the ±1 result is fp32, so the
+        # reconstructed weight w = (b1 @ b2).t() is fp32 and F.linear(x_bf16,
+        # w_fp32) raises a dtype-mismatch RuntimeError. Cast b back to x's dtype
+        # so the binarized matrix matches the (possibly low-precision) input.
+        b = torch.where(x >= 0, 1.0, -1.0).to(x.dtype)
         # STE: forward = sign, backward = identity
         return x + (b - x).detach()
